requirements.txt
azure-identity==1.17.1
azure-mgmt-resource==23.1.0
azure-mgmt-network==25.3.0
azure-mgmt-storage==21.2.1
azure-mgmt-keyvault==10.3.0
jinja2==3.1.4
cvss==3.1
requests==2.32.3
